Adding an optional 'label' property to base Vertice class.
This replaces the 'context' class that existed on a handful of edge and vertice types; given heavy use of 'context' in the Spider class this led to a lot of confusion. Classes that require a label (for example ones that use it in their key) can simply add a non-optional versino of the property to their DataInterface.